+++ ocs_ci/ocs/ui/page_objects/block_and_file.py
@@ -1,8 +1,10 @@
 import time
 
+from ocs_ci.framework import config
 from ocs_ci.ocs.ui.helpers_ui import format_locator, logger
 from ocs_ci.ocs.ui.page_objects.storage_system_details import StorageSystemDetails
 from ocs_ci.ocs.ui.workload_ui import PvcCapacityDeploymentList, compare_mem_usage
+from ocs_ci.ocs.ui.validation_ui import ValidationUI
 
 
 class BlockAndFile(StorageSystemDetails):
@@ -199,3 +201,35 @@ class BlockAndFile(StorageSystemDetails):
             self.validation_loc["locate_average_of_storage_consumption"]
         )
         return (get_est_days_from_element, get_avg_from_element)
+
+    def odf_storagesystems_consumption_trend(self):
+        """
+        Function to verify changes and validate elements on ODF storage consumption trend for ODF 4.17
+        This will navigate through below order
+        DataFoundation>>Storage>>storagecluster_storagesystem_details>>Block and File page
+        Further it looks for the Consumption trend card
+
+        Returns:
+            tuple: tpl_of_days_and_avg  ex: (Estimated days, Average)
+
+        """
+        ValidationUI().odf_console_plugin_check()
+        storage_systems_page = (
+            ValidationUI().nav_odf_default_page().nav_storage_systems_tab()
+        )
+        # PageNavigator().nav_odf_default_page().nav_storage_systems_tab()
+        storage_system_details = (
+            storage_systems_page.nav_storagecluster_storagesystem_details()
+        )
+        storage_system_details.nav_details_overview()
+        storage_system_details.nav_details_object()
+
+        if not config.ENV_DATA["mcg_only_deployment"]:
+            tpl_of_days_and_avg = (
+                storage_system_details.nav_block_and_file().get_estimated_days_from_consumption_trend()
+            )
+            return tpl_of_days_and_avg
+
+        else:
+            logger.error("No data available for MCG-only deployments.")
+            return None
